Adds /STACKS and /CODE words, changes EXIT to use Exit instruction
Exit instruction is redudant with Unnest, but we currently use Unnest to
mark the end of a function in several places.  This is probably a
temporary solution that allows EXIT to work without affecting existing
behavior.

In the future, we should either have an explicit function length in the
dictionary or